Adds Inventory Automation
In older version of OpenStack deployed with TripleO, the
tripleo-ansible-inventory script returns json and lacks the capability
to generate an asible inventory file in yaml.  This patch adds a script
that will accept the output of the inventory data in json and transform
it to the appropriate yaml.

The script accepts the json output as piped input.  To use the script,
run the following command:

tripleo-ansible-inventory | scripts/make_inventory_file > inventory.yaml